+++ internal/config/reverse_test.go
@@ -26,7 +26,7 @@ func TestValidateReverseOptions(t *testing.T) {
 
 	t.Log("Reverse, (invalid, abc)")
 	{
-		expectedErr := errors.New("Invalid Reverse Check ID (abc)")
+		expectedErr := errors.New("invalid Reverse Check ID (abc)")
 		viper.Set(KeyCheckBundleID, "abc")
 		err := validateReverseOptions()
 		if err == nil {
@@ -39,7 +39,7 @@ func TestValidateReverseOptions(t *testing.T) {
 
 	t.Log("Reverse, (invalid, /check_bundle/abc)")
 	{
-		expectedErr := errors.New("Invalid Reverse Check ID (/check_bundle/abc)")
+		expectedErr := errors.New("invalid Reverse Check ID (/check_bundle/abc)")
 		viper.Set(KeyCheckBundleID, "/check_bundle/abc")
 		err := validateReverseOptions()
 		if err == nil {
